refactor: extract sync consensus strategy
